requirements.txt
flask
flask_cors
mariadb
paho-mqtt